+++ NEWS.md
@@ -1,6 +1,6 @@
 # quanteda.textmodels 0.9.1
 
-* Adds `data_corpus_dailnoconf1991` and `data_corpus_irishbudget2010` moved from the **quanteda** package.
+* Adds `data_corpus_dailnoconf1991` and `data_corpus_irishbudget2010` corpus objects moved from the **quanteda** package.
 
 
 # quanteda.textmodels 0.9.0