+++ NEWS.md
@@ -1,3 +1,7 @@
+# quanteda.textmodels 0.9.2
+
+* Added a logistic regression classifier, `textmodel_lr()`.
+
 # quanteda.textmodels 0.9.1
 
 * Adds `data_corpus_dailnoconf1991` and `data_corpus_irishbudget2010` corpus objects moved from the **quanteda** package.